+++ CHANGELOG.rst
@@ -3,7 +3,23 @@ Changelog
 =========
 
 ------------------
-0.0.0 (2021-06-14)
+0.0.0 (2021-06-23)
 ------------------
 
-* First release on PyPI.
+* First release on PyPI
+
+------------------
+0.0.1 (2021-06-25)
+------------------
+
+* Refactor code into smaller classes and files
+* Update Section.deep_string()
+* Update readme/docs
+
+------------------
+0.0.2 (2021-06-26)
+------------------
+
+* Fix bug when using Section.leaves or Section.children
+* Add tests/test_indepth_usage.py
+* Update readme/docs